# Remove old 'Controls' field which was replaced two years ago