This saves memory and CPU time encoding and decoding.

We assume that no probes two years old are still running, and we don't need to reconstruct the data for saved historical reports because controls are not relevant to historical nodes.

Before (best of 5):
```
BenchmarkReportUnmarshal-2   	       1	1115642386 ns/op	180600552 B/op	 2451487 allocs/op
BenchmarkReportMerge-2   	       5	 319579062 ns/op	63458179 B/op	  474721 allocs/op
```

After:
```
BenchmarkReportUnmarshal-2   	       1	1017682911 ns/op	174734200 B/op	 2451483 allocs/op
BenchmarkReportMerge-2   	       5	 284404766 ns/op	61683891 B/op	  474442 allocs/op
```